requirements.txt
asgiref==3.2.10
cffi==1.14.0
cryptography==3.3.2
Django==3.1.8
django-filter==2.4.0
django-rest-knox==4.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.2
gunicorn==20.0.4
PlexAPI==4.6.0
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
youtube-dl==2020.11.24
